Move SendInvoice functions to a new file
