+++ src/pages/iou/request/step/IOURequestStepCompanyInfo.tsx
@@ -16,7 +16,8 @@ import {convertToDisplayString} from '@libs/CurrencyUtils';
 import {extractUrlDomain} from '@libs/Url';
 import {getFieldRequiredErrors, isPublicDomain, isValidWebsite} from '@libs/ValidationUtils';
 import Navigation from '@navigation/Navigation';
-import {getIOURequestPolicyID, sendInvoice} from '@userActions/IOU';
+import {getIOURequestPolicyID} from '@userActions/IOU';
+import {sendInvoice} from '@userActions/IOU/SendInvoice';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type SCREENS from '@src/SCREENS';
